# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitLog API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface PhotoUploadResponse {
  url: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Profile {
  id: number;
  userId: number;
  firstName: string;
  lastName: string;
  photoUrl?: string | null;
  age?: number | null;
  gender?: string | null;
  heightCm?: number | null;
  weightKg?: number | null;
  fitnessGoals: string[];
  activityLevel?: string | null;
  dailyCalorieGoal?: number | null;
  dailyProteinGoal?: number | null;
  dailyCarbsGoal?: number | null;
  dailyFatGoal?: number | null;
}

export interface UpdateProfileRequest {
  firstName: string;
  lastName: string;
  age?: number | null;
  gender?: string | null;
  heightCm?: number | null;
  weightKg?: number | null;
  fitnessGoals?: string[];
  activityLevel?: string | null;
  dailyCalorieGoal?: number | null;
  dailyProteinGoal?: number | null;
  dailyCarbsGoal?: number | null;
  dailyFatGoal?: number | null;
}

export interface WorkoutSet {
  id: number;
  exerciseId: number;
  reps?: number | null;
  weightKg?: number | null;
  order: number;
}

export interface WorkoutExercise {
  id: number;
  workoutId: number;
  name: string;
  order: number;
  sets: WorkoutSet[];
}

export type WorkoutMetadata = { [key: string]: unknown } | null;

export interface Workout {
  id: number;
  userId: number;
  activityType: string;
  name?: string | null;
  date: string;
  durationMinutes?: number | null;
  distanceKm?: number | null;
  caloriesBurned?: number | null;
  mood?: string | null;
  notes?: string | null;
  metadata?: WorkoutMetadata;
  exercises: WorkoutExercise[];
  createdAt: string;
}

export type CreateWorkoutRequestMetadata = { [key: string]: unknown } | null;

export interface CreateSetRequest {
  reps?: number | null;
  weightKg?: number | null;
  order: number;
}

export interface CreateExerciseRequest {
  name: string;
  order: number;
  sets: CreateSetRequest[];
}

export interface CreateWorkoutRequest {
  activityType: string;
  name?: string | null;
  date: string;
  durationMinutes?: number | null;
  distanceKm?: number | null;
  caloriesBurned?: number | null;
  mood?: string | null;
  notes?: string | null;
  metadata?: CreateWorkoutRequestMetadata;
  exercises?: CreateExerciseRequest[];
}

export interface WorkoutListResponse {
  workouts: Workout[];
  total: number;
}

export interface TodayStats {
  caloriesBurned: number;
  activeMinutes: number;
  workoutsCompleted: number;
  mealsLogged: number;
}

export interface DayStat {
  date: string;
  dayLabel: string;
  activeMinutes: number;
  isToday: boolean;
}

export interface WeeklyStats {
  days: DayStat[];
}

export interface WeekFrequency {
  weekLabel: string;
  count: number;
}

export interface ActivityBreakdown {
  activityType: string;
  count: number;
  percentage: number;
}

export interface WorkoutSummary {
  totalThisWeek: number;
  totalThisMonth: number;
  weeklyFrequency: WeekFrequency[];
  activityBreakdown: ActivityBreakdown[];
}

export interface RecentActivityItem {
  id: number;
  type: string;
  name: string;
  date: string;
  keyStat?: string | null;
  activityType?: string | null;
}

export interface RecentActivityResponse {
  activities: RecentActivityItem[];
}

export interface MealFoodItem {
  id: number;
  mealId: number;
  name: string;
  portionSize: number;
  unit: string;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}

export interface Meal {
  id: number;
  userId: number;
  name: string;
  category: string;
  date: string;
  photoUrl?: string | null;
  notes?: string | null;
  foodItems: MealFoodItem[];
  totalCalories: number;
  totalProteinG: number;
  totalCarbsG: number;
  totalFatG: number;
  createdAt: string;
}

export interface CreateFoodItemRequest {
  name: string;
  portionSize: number;
  unit: string;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}

export interface CreateMealRequest {
  name: string;
  category: string;
  date: string;
  photoUrl?: string | null;
  notes?: string | null;
  foodItems: CreateFoodItemRequest[];
}

export interface DailyNutritionTotals {
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}

export interface MealDayResponse {
  date: string;
  meals: Meal[];
  dailyTotals: DailyNutritionTotals;
  calorieGoal?: number | null;
}

export interface MacroSplit {
  proteinPercentage: number;
  carbsPercentage: number;
  fatPercentage: number;
}

export interface DailyCalories {
  date: string;
  calories: number;
  goal?: number | null;
}

export interface NutritionStats {
  avg7DayCalories: number;
  avg30DayCalories: number;
  macroSplit: MacroSplit;
  dailyCalories: DailyCalories[];
}

export interface Equipment {
  id: number;
  userId: number;
  name: string;
  category: string;
  photoUrl?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateEquipmentRequest {
  name: string;
  category: string;
  photoUrl?: string | null;
  notes?: string | null;
}

export interface EquipmentListResponse {
  equipment: Equipment[];
}

export interface BodyMeasurement {
  id: number;
  userId: number;
  date: string;
  weightKg?: number | null;
  bodyFatPercent?: number | null;
  chestCm?: number | null;
  waistCm?: number | null;
  hipsCm?: number | null;
  armsCm?: number | null;
  createdAt: string;
}

export interface CreateMeasurementRequest {
  date: string;
  weightKg?: number | null;
  bodyFatPercent?: number | null;
  chestCm?: number | null;
  waistCm?: number | null;
  hipsCm?: number | null;
  armsCm?: number | null;
}

export interface MeasurementListResponse {
  measurements: BodyMeasurement[];
}

export interface StreakResponse {
  currentWorkoutStreak: number;
  longestWorkoutStreak: number;
  currentMealStreak: number;
}

export interface PersonalRecord {
  label: string;
  value: string;
  date?: string | null;
  activityType?: string | null;
}

export interface PersonalRecordsResponse {
  records: PersonalRecord[];
}

export interface Settings {
  id: number;
  userId: number;
  darkMode: boolean;
  unitSystem: string;
  notificationsEnabled: boolean;
}

export interface UpdateSettingsRequest {
  darkMode?: boolean;
  unitSystem?: string;
  notificationsEnabled?: boolean;
}

export interface ExportDataResponse {
  workouts: Workout[];
  meals: Meal[];
  measurements: BodyMeasurement[];
}

export type UploadProfilePhotoBody = {
  photo?: Blob;
};

export type GetWorkoutsParams = {
  limit?: number;
  offset?: number;
  date?: string;
};

export type GetMealsParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date?: string;
};

export type GetMeasurementsParams = {
  /**
   * Number of days to look back (30, 90, 365)
   */
  days?: number;
};