# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitLog API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PROFILE
...
  /meals:
    get:
      operationId: getMeals
      tags: [meals]
      summary: Get meals for a date
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Meals for the date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealDayResponse"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Create a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealRequest"
      responses:
        "201":
          description: Meal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}:
    get:
      operationId: getMeal
      tags: [meals]
      summary: Get a specific meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Meal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    put:
      operationId: updateMeal
      tags: [meals]
      summary: Update a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealRequest"
      responses:
        "200":
          description: Meal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Meal deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /equipment:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get user equipment
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentListResponse"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Add equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentRequest"
      responses:
        "201":
          description: Equipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentRequest"
      responses:
        "200":
          description: Equipment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Equipment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # MEASUREMENTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
      required: [email, password, firstName, lastName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, firstName, lastName, createdAt]

    # PROFILE SCHEMAS
...
    MealFoodItem:
      type: object
      properties:
        id:
          type: number
        mealId:
          type: number
        name:
          type: string
        portionSize:
          type: number
        unit:
          type: string
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
      required: [id, mealId, name, portionSize, unit, calories, proteinG, carbsG, fatG]
    Meal:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        name:
          type: string
        category:
          type: string
        date:
          type: string
          format: date-time
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        foodItems:
          type: array
          items:
            $ref: "#/components/schemas/MealFoodItem"
        totalCalories:
          type: number
        totalProteinG:
          type: number
        totalCarbsG:
          type: number
        totalFatG:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, category, date, foodItems, totalCalories, totalProteinG, totalCarbsG, totalFatG, createdAt]
    CreateMealRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        date:
          type: string
          format: date-time
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        foodItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateFoodItemRequest"
      required: [name, category, date, foodItems]
    CreateFoodItemRequest:
      type: object
      properties:
        name:
          type: string
        portionSize:
          type: number
        unit:
          type: string
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
      required: [name, portionSize, unit, calories, proteinG, carbsG, fatG]
    MealDayResponse:
      type: object
      properties:
        date:
          type: string
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
        dailyTotals:
          $ref: "#/components/schemas/DailyNutritionTotals"
        calorieGoal:
          type: number
          nullable: true
      required: [date, meals, dailyTotals]
    DailyNutritionTotals:
      type: object
      properties:
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
      required: [calories, proteinG, carbsG, fatG]
...
    Equipment:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        name:
          type: string
        category:
          type: string
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, category, createdAt]
    CreateEquipmentRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        photoUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, category]
    EquipmentListResponse:
      type: object
      properties:
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/Equipment"
      required: [equipment]

    # MEASUREMENT SCHEMAS
...