# lib/api-zod/src/generated/types/mealDayResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitLog API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyNutritionTotals } from "./dailyNutritionTotals";
import type { Meal } from "./meal";

export interface MealDayResponse {
  date: string;
  meals: Meal[];
  dailyTotals: DailyNutritionTotals;
  calorieGoal?: number | null;
}